requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.7
django-filter==21.1
django-money==2.1
djangorestframework==3.12.4
Flask==2.0.1
httpie==2.5.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
legacy==0.1.6
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.3.2
py-moneyed==1.2
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.7
Werkzeug==2.0.1
